+++ packages/core/mashroom-json-schemas/type-definitions/mashroom-remote-package-scanner.d.ts
@@ -5,7 +5,7 @@
  * and run json-schema-to-typescript to regenerate this file.
  */
 
-export interface MashroomPortalRemoteApps {
+export interface MashroomRemotePackageScanner {
   $schema?: any;
-  remotePortalApps: string[];
+  remotePackageUrls: string[];
 }