+++ packages/plugin-packages/mashroom-remote-package-scanner/src/js/scanner/mashroom-bootstrap-plugin-package-scanner.ts
@@ -1,0 +1,22 @@
+
+import {registerRemoteAppMetrics, unregisterRemoteAppMetrics} from '../metrics/metrics';
+import healthProbe from '../health/health-probe';
+import RemotePluginPackagesScanner from './RemotePluginPackagesScanner';
+import type {MashroomPluginPackageScannerPluginBootstrapFunction} from '@mashroom/mashroom/type-definitions';
+
+const bootstrap: MashroomPluginPackageScannerPluginBootstrapFunction = async (pluginName, pluginConfig, pluginContextHolder) => {
+    const { remotePackageUrls } = pluginConfig;
+    const {services: {core: {pluginService, healthProbeService}}, serverConfig: {serverRootFolder}} = pluginContextHolder.getPluginContext();
+
+    healthProbeService.registerProbe(pluginName, healthProbe(pluginContextHolder));
+    registerRemoteAppMetrics(pluginContextHolder);
+
+    pluginService.onUnloadOnce(pluginName, () => {
+        healthProbeService.unregisterProbe(pluginName);
+        unregisterRemoteAppMetrics();
+    });
+
+    return new RemotePluginPackagesScanner(remotePackageUrls, serverRootFolder, pluginContextHolder);
+};
+
+export default bootstrap;