+++ packages/plugin-packages/mashroom-remote-package-scanner/src/js/webapp/mashroom-bootstrap-webapp.ts
@@ -1,0 +1,16 @@
+
+import context from '../context';
+import webapp from './webapp';
+
+import type {
+    MashroomWebAppPluginBootstrapFunction
+} from '@mashroom/mashroom/type-definitions';
+
+const bootstrap: MashroomWebAppPluginBootstrapFunction = async (pluginName, pluginConfig) => {
+    const { showAddRemoteAppForm } = pluginConfig;
+    context.webUIShowAddRemoteAppForm = showAddRemoteAppForm;
+
+    return webapp;
+};
+
+export default bootstrap;